pyproject.toml
[tool.poetry]
name = "cryptomarket"
version = "0.1.0"
description = ""
authors = ["frozjeee <miras-d@bk.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
aio-pika = "^9.0.5"
alembic = "^1.10.2"
ruff = "^0.0.259"
pycryptodome = "^3.17"
pyllist = "^0.3"
sqlalchemy = "1.4.46"
idanalyzer = "^1.2.2"
python-dotenv = "^1.0.0"
asyncpg = "^0.27.0"
pytz = "^2023.3"
jinja2 = "^3.1.2"
uvicorn = "^0.21.1"
fastapi = "^0.95.1"
passlib = "^1.7.4"
jwt = "^1.3.1"
python-jose = "^3.3.0"
pydantic = {extras = ["email"], version = "^1.10.7"}
python-multipart = "^0.0.6"
psycopg2 = "^2.9.6"
bcrypt = "^4.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.ruff]
line-length = 120
